Enforce a delay between batches
When `uploadBatch` is called from the loop, all batches' first requests
are fired at the exact same time, kind of defeating the object somewhat.

This replaces the loop with a recursive call to `uploadBatch` from the
`loadend` event, which ensures the batch has actually completed before
the request for the next one can be fired.

Also adds a delay between batches and when moving completed items which
significantly improves the overall visual feel of the UI when uploading.