Fix issue when logs get processed first
Sometimes when you drag a folder, entries may get processed in an order
such that a log file is added before the parent section exists for the
contact to which it belongs, resulting in an error in the console.

This commit fixes the issue by handling what would normally be done in
the `addFolder` function if for example the photos folder was first.