Improve header injection handling
Moves the adding of headers into the main drop handler so that we have
better control over the position, so the global files can come first.

Also adjusts subheadings to have the same (left) padding as tabs, which
is reduced a little from the px-7 value used in this commit previously.