Disable pending/complete tabs when empty
This disables the Completed tab until files in that section are being
prepared for upload, at which point the `pointer-events-none` class is
removed. Once the final file in a section is uploaded, that class will
then be added instead to the Pending tab so you can't switch back to it.